refactor: 공통으로 사용되는 type을 분리 (types/common.ts)
