+++ src/components/templates/generateMovieItems.ts
@@ -1,4 +1,4 @@
-import { HTMLTemplate } from "../abstract/BaseComponent";
+import { HTMLTemplate } from "../../types/common";
 import IMAGES from "../../images";
 
 export type Path = string;
@@ -16,22 +16,22 @@ const generateMovieItem = (movie: Movie): HTMLTemplate => {
   const { title, posterPath, voteAverage } = movie;
 
   return `
-  <li>
-    <a href="#">
-      <div class="item-card">
-        <img class="item-thumbnail" src="${BASE_POSTER_URL}${posterPath}" loading="lazy" alt="${title}" />
-        <p class="item-title">${title}</p>
-        <p class="item-score">
-          <img src="${IMAGES.starFilled}" alt="별점" />
-          ${voteAverage}
-        </p>
-      </div>
-    </a>
-  </li>
+    <li>
+      <a href="#">
+        <div class="item-card">
+          <img class="item-thumbnail" src="${BASE_POSTER_URL}${posterPath}" loading="lazy" alt="${title}" />
+          <p class="item-title">${title}</p>
+          <p class="item-score">
+            <img src="${IMAGES.starFilled}" alt="별점" />
+            ${voteAverage}
+          </p>
+        </div>
+      </a>
+    </li>
   `;
 };
 
-export const generateMovieItems = (movies: Movie[]) => {
+export const generateMovieItems = (movies: Movie[]): HTMLTemplate => {
   return movies.reduce((movieListTemplate, movieInfo) => {
     return movieListTemplate + generateMovieItem(movieInfo);
   }, "");
