+++ src/types/common.ts
@@ -1,0 +1,4 @@
+export type HTMLTemplate = string;
+export type TargetId = string;
+export type Path = string;
+export type Query = string;
